requirements.txt
customtkinter
pillow
opencv-python
scikit-learn
numpy
